Rename: 폴더 이름 변경 - 프로필 수정 페이지 폴더 이름 변경 기존 ProfileSetting -> ProfileEdit - 회원가입 후 설정하는 프로필 페이지를 ProfileSetting으로 사용할 예정
