+++ src/pages/Profile/ProfileEdit/index.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function ProfileEdit() {
+  return <div>ProfileEdit</div>;
+}
+
+export default ProfileEdit;
